upgrade boilerplate, fix path for external
