+++ .eslintrc.js
@@ -17,6 +17,5 @@ module.exports = {
     'import/extensions': ['error', {
       js: 'always',
     }],
-    'function-paren-newline': 'off',
   },
 };
