+++ README.md
@@ -5,10 +5,22 @@ This project is for the https://main--walgreens--hlxsites.hlx.page/ website.
 - Preview: https://main--walgreens--hlxsites.hlx.page/
 - Live: https://main--walgreens--hlxsites.hlx.live/
 
+## Installation
+
+```sh
+npm i
+```
+
+## Linting
+
+```sh
+npm run lint
+```
+
 ## Local development
 
-1. Create a new repository based on the `helix-project-boilerplate` template and add a mountpoint in the `fstab.yaml`
-1. Add the [helix-bot](https://github.com/apps/helix-bot) to the repository
-1. Install the [Helix CLI](https://github.com/adobe/aem-cli): `npm install -g @adobe/aem-cli`
-1. Start Franklin Proxy: `aem up` (opens your browser at `http://localhost:3000`)
+1. Create a new repository based on the `aem-boilerplate` template and add a mountpoint in the `fstab.yaml`
+1. Add the [AEM Code Sync GitHub App](https://github.com/apps/aem-code-sync) to the repository
+1. Install the [AEM CLI](https://github.com/adobe/aem-cli): `npm install -g @adobe/aem-cli`
+1. Start AEM Proxy: `aem up` (opens your browser at `http://localhost:3000`)
 1. Open the `{repo}` directory in your favorite IDE and start coding :)
