+++ external-topic/blocks/breadcrumb/breadcrumb.css
@@ -37,7 +37,7 @@
   display: none;
 }
 
-.breadcrumb li a svg {
+.breadcrumb li a img {
   fill: #707575;
   bottom: -2px;
   height: 30px;
