+++ external-topic/blocks/carousel/carousel.css
@@ -62,12 +62,15 @@
   z-index: 2;
 }
 
+.carousel .carousel-nav button.carousel-nav-left {
+}
+
 .carousel .icon-arrow-right {
   background-color: var(--highlight-background-color);
   border-radius: 50%;
 }
 
-.carousel .carousel-nav .icon svg {
+.carousel .carousel-nav .icon img {
   fill: #989090;
   width: 30px;
   height: 32px;
@@ -92,8 +95,8 @@
     width: 30px;
   }
 
-  .carousel .carousel-nav .icon svg {
-    fill: var(--background-color);
+  .carousel .carousel-nav .icon img {
+    filter: brightness(0) invert(1);
   }
 
   .carousel .carousel-nav button.disabled {
