+++ head.html
@@ -1,5 +1,5 @@
 <meta name="viewport" content="width=device-width, initial-scale=1" />
-<script src="/external-topic/scripts/lib-franklin.js" type="module"></script>
+<script src="/external-topic/scripts/aem.js" type="module"></script>
 <script src="/external-topic/scripts/scripts.js" type="module"></script>
 <link rel="stylesheet" href="/external-topic/styles/styles.css" />
 <link rel="preload" href="/external-topic/external-styles/header-clientCSSContent.css" as="style"/>
